Update README.md
This is an application of the Finite Elements Method.
In this example I make use of quadrilateral finite elements in order to find the displacement field of a rectangular structural beam. The size of the structure is reconfigurable as well as 
the density of the meshing, the loading conditions (magnitude and topology of the force field) and the boundary conditions.

An extention of the current code is under developement and I will shortly upload it. The extention performs a Monte Carlo simulation assuming that the modulus of elasticity is a stochastic field.
To generate the realizations of this stochastic field we are using the Karhunen-Loeve series expansion.

Feel free to run the code or even better suggest improvements.